requirements.txt
ccxt
openai
pandas
schedule
python-dotenv
requests
urllib3
python-telegram-bot